+++ dmscreen/gui/application.py
@@ -4,8 +4,8 @@ import gi
 gi.require_version('Gtk', '3.0')
 from gi.repository import Gio, Gtk  # noqa: E402
 
-from dmscreen.gui.pages.characterpage import CharacterPage  # noqa: E402
-from dmscreen.gui.pages.homepage import HomePage  # noqa: E402
+from dmscreen.gui.pages.character_page import CharacterPage  # noqa: E402
+from dmscreen.gui.pages.home_page import HomePage  # noqa: E402
 from dmscreen.data.jsonconvert import ParseData  # noqa: E402
 
 UI_STRING = """
