+++ dmscreen/gui/elements/spells_sheet.py
@@ -1,5 +1,5 @@
 from gi.repository import Gtk
-from dmscreen.data.dataLoader import get_spell_names
+from dmscreen.data.data_loader import get_spell_names
 
 
 class SpellsSheet:
